// packages/beam/src/types.ts
export type TextureSource =
  | ImageBitmap
  | HTMLImageElement
  | HTMLCanvasElement
  | HTMLVideoElement
  | { data: BufferSource; width: number; height: number }
export type CubeFaces = [
  TextureSource,
  TextureSource,
  TextureSource,
  TextureSource,
  TextureSource,
  TextureSource,
]

export type Wrap = 'repeat' | 'mirror' | 'clamp'
export type Filter = 'nearest' | 'linear'

export interface TextureOpts {
  format?: GPUTextureFormat
  srgb?: boolean
  flipY?: boolean
  mips?: boolean
  usage?: GPUTextureUsageFlags
  label?: string
}
// ...
export interface Texture {
  readonly kind: 'texture'
  set(source: TextureSource | CubeFaces, opts?: TextureOpts): this
  readonly gpu: GPUTexture
  readonly view: GPUTextureView
  readonly cube: boolean
  destroy(): void
}